+++ src/components/permanentDashboardSection/alerts/AlertForm.js
@@ -1,0 +1,8 @@
+// import React, { useContext, useEffect, useState } from "react";
+// import { basePatientDetailsContext } from "../permanentDashboardProvider";
+// // import "./??.css";
+// import { useHistory, useParams } from 'react-router-dom';
+
+// export const AlertForm = () => {
+//     console.log("alert form")
+// }
