+++ src/components/permanentDashboardSection/alerts/AlertList.js
@@ -1,28 +1,41 @@
-import React, { useContext, useEffect } from "react";
-import { basePatientDetailsContext } from "../permanentDashboardProvider"
-import { AlertCard } from "./AlertCard";
+// import React, { useContext, useEffect } from "react";
+// import { basePatientDetailsContext } from "../permanentDashboardProvider"
+// import { useHistory } from "react-router-dom";
+// import { Card, Button } from "react-bootstrap";
 
 
-export const AlertList = () => {
-  const { basePatientDetails, getBasePatientDetails } = useContext(basePatientDetailsContext)
+// export const AlertList = () => {
+//   const { basePatientDetails, getBasePatientDetailsById } = useContext(basePatientDetailsContext)
+//   const userId = parseInt(localStorage.getItem("app_user"))
+//   const history = useHistory()
 
 
-  useEffect(() => {
-    getBasePatientDetails()
-  }, [])
+//   useEffect(() => {
+//     getBasePatientDetailsById(userId)
+//   }, [])
 
 
-  const userId = parseInt(localStorage.getItem("app_user"))
-
-
-  return (
-    <div className="alerts">
-      {basePatientDetails.map(basePatientDetail => {
-        if (userId === basePatientDetail.id) {
-          return <AlertCard key={basePatientDetail.id} //argument
-            basePatientDetail={basePatientDetail} /> //argument
-        }
-      })}
-    </div>
-  )
-};
+//   return (
+//     <div className="alerts">
+//       {basePatientDetails.hasOwnProperty("id") ?
+//       <Card style={{ width: '18rem' }}>
+//                 <Card.Body>
+//                     <Card.Title></Card.Title>
+//                     <Card.Subtitle className="mb-2 text-muted">Conditions</Card.Subtitle>
+//                     <Card.Text>{basePatientDetail.conditions}</Card.Text>
+//                     <Card.Subtitle className="mb-2 text-muted">Allergies</Card.Subtitle>
+//                     <Card.Text>{basePatientDetail.allergies}</Card.Text>
+//                 </Card.Body>
+//             </Card>
+//             :
+//             <Card style={{ width: '18rem' }}>
+//             <Card.Body>
+//               <Button onClick={() => { history.push("/alertsForm") }}>
+//                 Add
+//                       </Button>
+//             </Card.Body>
+//           </Card>
+// }
+//     </div>
+//   )
+// };
